# app/config/api/openapi.yml
openapi: 3.0.2
info:
  title: Forester API
  version: latest
paths:
...
  "/frame/delete":
    post:
      summary: Delete Frame
      description: Delete Frame
      parameters:
        - in: query
          name: id
          required: true
          schema:
            type: number
          description: Specific frame identifier
      responses:
        '301':
          description: OK
        '500':
          $ref: '#/components/responses/Error'
...
components:
  schemas:
    Error:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string
...
  responses:
...
    Error:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...